backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
neo4j==5.23.1
python-socketio==5.11.4
python-multipart==0.0.9
email-validator==2.1.1
